split into libgbd (implementation) + gbd (frontend), build libgbd .a and .so
