+++ include/libgbd/gbd.h
