+++ include/libgbd/rdram.h
