+++ src/libgbd/macros.h
