+++ src/libgbd/obstack.h
