+++ src/libgbd/vector.c
