+++ src/libgbd/vector.h
