+++ src/libgbd/vt.h
