+++ src/libgbd/warnings_errors.h
